api/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
google-api-core==1.33.2
google-api-python-client==2.64.0
cloud-sql-python-connector==0.9.3
pg8000==1.29.3
python-dotenv==0.20.0
